Remove old LayerBase implementation.
