+++ packages/office-ui-fabric-react/src/components/Layer/Layer.base.tsx
@@ -1,53 +1,30 @@
-/* tslint:disable:no-unused-variable */
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-/* tslint:enable:no-unused-variable */
 
 import { Fabric } from '../../Fabric';
 import { ILayerProps, ILayerStyleProps, ILayerStyles } from './Layer.types';
-import { BaseComponent, classNamesFunction, getDocument, setVirtualParent, createRef } from '../../Utilities';
 import {
-  getDefaultTarget,
-  registerLayer,
-  unregisterLayer,
-  setDefaultTarget,
-  notifyHostChanged
-} from './Layer.notification';
+  BaseComponent,
+  classNamesFunction,
+  customizable,
+  getDocument,
+  createRef,
+  setVirtualParent
+} from '../../Utilities';
+import { registerLayer, getDefaultTarget, unregisterLayer } from './Layer.notification';
 
 const getClassNames = classNamesFunction<ILayerStyleProps, ILayerStyles>();
 
+@customizable('Layer', ['theme', 'hostId'])
 export class LayerBase extends BaseComponent<ILayerProps, {}> {
   public static defaultProps: ILayerProps = {
     onLayerDidMount: () => undefined,
     onLayerWillUnmount: () => undefined
   };
 
-  private _rootElement = createRef<HTMLDivElement>();
   private _host: Node;
   private _layerElement: HTMLElement | undefined;
-  private _hasMounted: boolean;
-
-  /**
-   * Used for notifying applicable Layers that a host is available/unavailable and to re-evaluate Layers that
-   * care about the specific host.
-   * @deprecated
-   */
-  public static notifyHostChanged(id: string) {
-    notifyHostChanged(id);
-  }
-
-  /**
-   * Sets the default target selector to use when determining the host in which
-   * Layered content will be injected into. If not provided, an element will be
-   * created at the end of the document body.
-   *
-   * Passing in a falsey value will clear the default target and reset back to
-   * using a created element at the end of document body.
-   * @deprecated
-   */
-  public static setDefaultTarget(selector?: string) {
-    setDefaultTarget(selector);
-  }
+  private _rootElement = createRef<HTMLDivElement>();
 
   constructor(props: ILayerProps) {
     super(props);
@@ -61,27 +38,128 @@ export class LayerBase extends BaseComponent<ILayerProps, {}> {
     }
   }
 
+  public componentWillMount(): void {
+    this._layerElement = this._getLayerElement();
+  }
+
+  public componentWillUpdate(): void {
+    if (!this._layerElement) {
+      this._layerElement = this._getLayerElement();
+    }
+  }
+
   public componentDidMount(): void {
-    this.componentDidUpdate();
+    this._setVirtualParent();
+
+    const { onLayerDidMount, onLayerMounted } = this.props;
+    if (onLayerMounted) {
+      onLayerMounted();
+    }
+
+    if (onLayerDidMount) {
+      onLayerDidMount();
+    }
   }
 
   public componentWillUnmount(): void {
     this._removeLayerElement();
 
-    if (this.props.hostId) {
-      unregisterLayer(this.props.hostId, this);
+    const { onLayerWillUnmount, hostId } = this.props;
+    if (onLayerWillUnmount) {
+      onLayerWillUnmount();
+    }
+
+    if (hostId) {
+      unregisterLayer(hostId, this);
     }
   }
 
   public componentDidUpdate(): void {
-    const host = this._getHost();
+    this._setVirtualParent();
+  }
 
+  public render(): React.ReactNode {
+    const classNames = this._getClassNames();
+    const { eventBubblingEnabled } = this.props;
+
+    return (
+      <span className="ms-layer" ref={this._rootElement}>
+        {this._layerElement &&
+          ReactDOM.createPortal(
+            eventBubblingEnabled ? (
+              <Fabric className={classNames.content}>{this.props.children}</Fabric>
+            ) : (
+              <Fabric
+                className={classNames.content}
+                onClick={this._filterEvent}
+                onContextMenu={this._filterEvent}
+                onDoubleClick={this._filterEvent}
+                onDrag={this._filterEvent}
+                onDragEnd={this._filterEvent}
+                onDragEnter={this._filterEvent}
+                onDragExit={this._filterEvent}
+                onDragLeave={this._filterEvent}
+                onDragOver={this._filterEvent}
+                onDragStart={this._filterEvent}
+                onDrop={this._filterEvent}
+                onMouseDown={this._filterEvent}
+                onMouseEnter={this._filterEvent}
+                onMouseLeave={this._filterEvent}
+                onMouseMove={this._filterEvent}
+                onMouseOver={this._filterEvent}
+                onMouseOut={this._filterEvent}
+                onMouseUp={this._filterEvent}
+                onKeyDown={this._filterEvent}
+                onKeyPress={this._filterEvent}
+                onKeyUp={this._filterEvent}
+                onFocus={this._filterEvent}
+                onBlur={this._filterEvent}
+                onChange={this._filterEvent}
+                onInput={this._filterEvent}
+                onInvalid={this._filterEvent}
+                onSubmit={this._filterEvent}
+              >
+                {this.props.children}
+              </Fabric>
+            ),
+            this._layerElement
+          )}
+      </span>
+    );
+  }
+
+  /**
+   * Helper to stop events from bubbling up out of Layer.
+   */
+  private _filterEvent = (ev: React.SyntheticEvent<HTMLElement>): void => {
+    // We should just be able to check ev.bubble here and only stop events that are bubbling up. However, even though mouseenter and
+    //    mouseleave do NOT bubble up, they are showing up as bubbling. Therefore we stop events based on event name rather than ev.bubble.
+    if (ev.type !== 'mouseenter' && ev.type !== 'mouseleave') {
+      ev.stopPropagation();
+    }
+  };
+
+  private _getClassNames() {
     const { className, styles, theme } = this.props;
     const classNames = getClassNames(styles!, {
       theme: theme!,
       className,
       isNotHost: !this.props.hostId
     });
+
+    return classNames;
+  }
+
+  private _setVirtualParent() {
+    if (this._rootElement && this._rootElement.current && this._layerElement) {
+      setVirtualParent(this._layerElement, this._rootElement.current);
+    }
+  }
+
+  private _getLayerElement(): HTMLElement | undefined {
+    const host = this._getHost();
+
+    const classNames = this._getClassNames();
 
     if (host !== this._host) {
       this._removeLayerElement();
@@ -91,10 +169,8 @@ export class LayerBase extends BaseComponent<ILayerProps, {}> {
       this._host = host;
 
       if (!this._layerElement) {
-        const rootElement = this._rootElement.current;
-        const doc = getDocument(rootElement);
-
-        if (!doc || !rootElement) {
+        const doc = getDocument();
+        if (!doc) {
           return;
         }
 
@@ -102,52 +178,28 @@ export class LayerBase extends BaseComponent<ILayerProps, {}> {
         this._layerElement.className = classNames.root!;
 
         host.appendChild(this._layerElement);
-        setVirtualParent(this._layerElement, rootElement);
       }
-
-      // Using this 'unstable' method allows us to retain the React context across the layer projection.
-      ReactDOM.unstable_renderSubtreeIntoContainer(
-        this,
-        <Fabric className={classNames.content}>{this.props.children}</Fabric>,
-        this._layerElement,
-        () => {
-          if (!this._hasMounted) {
-            this._hasMounted = true;
-
-            // TODO: @deprecated cleanup required.
-            if (this.props.onLayerMounted) {
-              this.props.onLayerMounted();
-            }
-
-            this.props.onLayerDidMount!();
-          }
-        }
-      );
     }
-  }
 
-  public render(): JSX.Element {
-    return <span className="ms-Layer" ref={this._rootElement} />;
+    return this._layerElement;
   }
 
   private _removeLayerElement(): void {
     if (this._layerElement) {
       this.props.onLayerWillUnmount!();
 
-      ReactDOM.unmountComponentAtNode(this._layerElement);
       const parentNode = this._layerElement.parentNode;
       if (parentNode) {
         parentNode.removeChild(this._layerElement);
       }
       this._layerElement = undefined;
-      this._hasMounted = false;
     }
   }
 
   private _getHost(): Node | undefined {
     const { hostId } = this.props;
-    const doc = getDocument(this._rootElement.current);
 
+    const doc = getDocument();
     if (!doc) {
       return undefined;
     }
